+++ agent-access-monitor/scripts/private/Get-AAMValidationResults.ps1
@@ -209,7 +209,7 @@ function Get-AAMValidationResults {
             # Optional zone filter (not applied when 'All')
             if ($Zone -ne 'All') {
                 # Map zone name to Dataverse option set integer
-                $zoneIntMap = @{ 'Zone1' = 100000001; 'Zone2' = 100000002; 'Zone3' = 100000003 }
+                $zoneIntMap = @{ '1' = 100000001; '2' = 100000002; '3' = 100000003 }
                 $zoneInt = if ($zoneIntMap.ContainsKey($Zone)) { $zoneIntMap[$Zone] } else { $Zone }
                 $violationFilters += "fsi_zone eq $zoneInt"
             }